Move page title from root into folder
