+++ app/src/components/PageTitle.tsx
@@ -26,7 +26,6 @@ const PageTitle = ({ title, favicon }: Props) => {
 					sizes="32x32"
 				/>
 			)}
-			<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		</Helmet>
 	);
 };
